Reduce warnings

+++ include/ruby/onigmo.h
@@ -744,8 +744,8 @@ typedef struct {
 typedef struct {
   int lower;
   int upper;
-  int base_num;
-  int inner_num;
+  long base_num;
+  long inner_num;
 } OnigRepeatRange;
 
 typedef void (*OnigWarnFunc)(const char* s);
